Extract deprecations for readability
This prepares deprecations for adding a new deprecation to handle the addition of the `verticalAlignment` attribute.
